+++ benchmark/benchmark_geofileops.py
@@ -1,14 +1,10 @@
-"""
-Run the benchmarks specified.
-"""
+"""Run the benchmarks specified."""
 
 from benchmark import benchmarker
 
 
 def main():
-    """
-    Run the benchmark function(s).
-    """
+    """Run the benchmark function(s)."""
     functions_to_run = [
         "export_by_location_intersects_complexpoly",
         # "export_by_location_intersects",