+++ CHANGELOG.md
@@ -35,6 +35,7 @@
   raised. (#518)
 - Small changes to support geopandas 1.0 (#529)
 - Linting improvements: ruff version to 0.4.4 + apply import sorting (#531)
+- Add support for renaming column with only difference in casing (#549)
 
 ## 0.8.2 (2024-05-25)